+++ payshield/__init__.py
@@ -1,0 +1,32 @@
+# payShield test utility by Marco S. Zuppone - msz@msz.eu
+# Project name: payShieldPressureTest
+# Official GitHub Repository: https://github.com/mszeu/PayShieldPressureTest
+# Copyright (C) 2020-2026 by Marco S. Zuppone - msz@msz.eu
+#
+# AGPL-3.0 License — see LICENSE file for details.
+
+"""
+payshield
+---------
+Core package for payShieldPressureTest.
+
+Public re-exports so callers can do, e.g.:
+    from payshield import PayConnector, run_test
+"""
+
+from .connector import PayConnector
+from .decoders import DECODERS
+from .error_codes import payshield_error_codes, PAYSHIELD_ERROR_CODE
+from .runner import run_test, check_return_message, check_returned_command_verb
+from .updater import UpdateChecker
+
+__all__ = [
+    "PayConnector",
+    "DECODERS",
+    "payshield_error_codes",
+    "PAYSHIELD_ERROR_CODE",
+    "run_test",
+    "check_return_message",
+    "check_returned_command_verb",
+    "UpdateChecker",
+]
